Refactor backing away into its own goal
