+++ src/main/java/org/minefortress/entity/ArcherPawn.java
@@ -17,9 +17,7 @@ import net.remmintan.mods.minefortress.core.interfaces.entities.pawns.IProfessio
 import net.remmintan.mods.minefortress.core.interfaces.entities.pawns.IWarrior;
 import org.minefortress.entity.ai.goal.EatGoal;
 import org.minefortress.entity.ai.goal.SelectTargetToAttackGoal;
-import org.minefortress.entity.ai.goal.warrior.FollowLivingEntityGoal;
-import org.minefortress.entity.ai.goal.warrior.MoveToBlockGoal;
-import org.minefortress.entity.ai.goal.warrior.RangedAttackGoal;
+import org.minefortress.entity.ai.goal.warrior.*;
 
 public class ArcherPawn extends TargetedPawn implements IWarrior, RangedAttackMob, IProfessional {
 
@@ -31,6 +29,7 @@ public class ArcherPawn extends TargetedPawn implements IWarrior, RangedAttackMo
     protected void initGoals() {
         super.initGoals();
         this.goalSelector.add(1, new RangedAttackGoal(this));
+        this.goalSelector.add(1, new BackAwayGoal(this, 8.0));
         this.goalSelector.add(2, new MoveToBlockGoal(this));
         this.goalSelector.add(2, new FollowLivingEntityGoal(this));
         this.goalSelector.add(3, new EatGoal(this));
